Not finished with testing yet.
Fix: I fixed the error messages I encountered while starting the tests. It was a mistake of deleted fields that I hadn't removed from all the places.
To fix: When testing via Postman, I encountered an issue where I need a piece of authorization. So, I'm going to create that and switch to another branch.